Remove code duplicates removed mobile order-row added mobile row rendering to base order-row
