+++ shared/pages/Home/Orders/Orders.js
@@ -4,8 +4,6 @@ import { connect } from 'redaction'
 import actions from 'redux/actions'
 import { withRouter } from 'react-router-dom'
 
-import { isMobile } from 'react-device-detect'
-
 import constants from 'helpers/constants'
 
 import cssModules from 'react-css-modules'
@@ -19,7 +17,6 @@ import PageSeo from 'components/Seo/PageSeo'
 
 import Pair from './Pair'
 import Row from './Row/Row'
-import RowMobile from './RowMobile/RowMobile'
 import MyOrders from './MyOrders/MyOrders'
 import { FormattedMessage } from 'react-intl'
 
@@ -156,11 +153,7 @@ export default class Orders extends Component {
           titles={titles}
           rows={sellOrders}
           rowRender={(row, index) => (
-            isMobile &&  <RowMobile
-              key={index}
-              orderId={orderId}
-              row={row}
-            /> || <Row
+            <Row
               key={index}
               orderId={orderId}
               row={row}
@@ -187,11 +180,7 @@ export default class Orders extends Component {
           titles={titles}
           rows={buyOrders}
           rowRender={(row, index) => (
-            isMobile &&  <RowMobile
-              key={index}
-              orderId={orderId}
-              row={row}
-            /> || <Row
+            <Row
               key={index}
               orderId={orderId}
               row={row}
