+++ shared/pages/Home/Orders/Row/Row.js
@@ -4,6 +4,11 @@ import PropTypes from 'prop-types'
 import { connect } from 'redaction'
 import actions from 'redux/actions'
 
+import { isMobile } from 'react-device-detect'
+
+import cssModules from 'react-css-modules'
+import styles from './Row.scss'
+
 import { links, constants } from 'helpers'
 import { Link, Redirect } from 'react-router-dom'
 
@@ -20,6 +25,7 @@ import { FormattedMessage } from 'react-intl'
 @connect({
   peer: 'ipfs.peer',
 })
+@cssModules(styles)
 export default class Row extends Component {
 
   static propTypes = {
@@ -87,12 +93,23 @@ export default class Row extends Component {
 
     actions.core.updateCore()
   }
-
-  render() {
+  renderWebContent() {
     const { balance, isFetching } = this.state
-    const { orderId, row: { id, buyCurrency, sellCurrency, isMy, buyAmount,
-      sellAmount, isRequested, isProcessing,
-      owner: {  peer: ownerPeer } }, peer } = this.props
+    const {
+      orderId,
+      row: {
+        id,
+        buyCurrency,
+        sellCurrency,
+        isMy,
+        buyAmount,
+        sellAmount,
+        isRequested,
+        isProcessing,
+        owner: {  peer: ownerPeer }
+      },
+      peer
+    } = this.props
 
     const pair = Pair.fromOrder(this.props.row)
 
@@ -194,5 +211,121 @@ export default class Row extends Component {
         </td>
       </tr>
     )
+  }
+  renderMobileContent() {
+    const { balance, isFetching } = this.state
+    const {
+      orderId,
+      row: {
+        id,
+        buyCurrency,
+        sellCurrency,
+        isMy,
+        buyAmount,
+        sellAmount,
+        isRequested,
+        isProcessing,
+        owner: {  peer: ownerPeer }
+      },
+      peer
+    } = this.props
+
+    const pair = Pair.fromOrder(this.props.row)
+
+    const { price, amount, total, main, base, type } = pair
+
+    return (
+      <tr
+        className="mobile-row"
+        style={orderId === id ? { background: 'rgba(0, 236, 0, 0.1)' } : {}}
+      >
+        <td>
+          <div styleName="bigContainer">
+            <div styleName="tdContainer-1">
+              <span styleName="firstType">
+                {type === PAIR_TYPES.BID ? 'You have' : 'You get'}
+              </span>
+              <span>{`${amount.toFixed(5)} ${main}`}</span>
+            </div>
+            <div><i className="fas fa-exchange-alt" /></div>
+            <div styleName="tdContainer-2">
+              <span styleName="secondType">
+                <FormattedMessage id="RowM122" defaultMessage="You get" />
+              </span>
+              <span>{`${total.toFixed(5)} ${base}`}</span>
+            </div>
+          </div>
+        </td>
+        <td>
+          {
+            peer === ownerPeer ? (
+              <RemoveButton onClick={() => this.removeOrder(id)} />
+            ) : (
+              <Fragment>
+                {
+                  isRequested ? (
+                    <Fragment>
+                      <div style={{ color: 'red' }}>
+                        <FormattedMessage id="RowM136" defaultMessage="REQUESTING" />
+                      </div>
+                      <Link to={`${links.swap}/${buyCurrency}-${sellCurrency}/${id}`}>
+                        <FormattedMessage id="RowM139" defaultMessage="Go to the swap" />
+                      </Link>
+                    </Fragment>
+                  ) : (
+                    isProcessing ? (
+                      <span>
+                        <FormattedMessage id="RowM145" defaultMessage="This order is in execution" />
+                      </span>
+                    ) : (
+                      isFetching ? (
+                        <Fragment>
+                          <InlineLoader />
+                          <br />
+                          <span>
+                            <FormattedMessage id="RowM153" defaultMessage="Please wait while we confirm your request" />
+                          </span>
+                        </Fragment>
+                      ) : (
+                        <RequestButton
+                          styleName="startButton"
+                          disabled={balance >= Number(buyAmount)}
+                          onClick={() => this.sendRequest(id, isMy ? sellCurrency : buyCurrency)}
+                          data={{ type, amount, main, total, base }}
+                          onMouseEnter={() => this.setState(() => ({ enterButton: true }))}
+                          onMouseLeave={() => this.setState(() => ({ enterButton: false }))}
+                          move={this.state.enterButton}
+                        >
+                          <FormattedMessage id="RowM166" defaultMessage="Start" />
+                        </RequestButton>
+                      )
+                    )
+                  )
+                }
+              </Fragment>
+            )
+          }
+        </td>
+      </tr>
+    )
+  }
+  render() {
+    const {
+      row: {
+        id,
+        buyCurrency,
+        sellCurrency,
+      },
+    } = this.props
+
+    if (this.state.redirect) {
+      return <Redirect push to={`${links.swap}/${buyCurrency}-${sellCurrency}/${id}`} />
+    }
+
+    if (isMobile) {
+      return this.renderMobileContent();
+    } else {
+      return this.renderWebContent();
+    }
   }
 }
