--- shared/pages/Home/Orders/RowMobile/RowMobile.scss
@@ -1,27 +1,0 @@
-.firstType,
-.secondType {
-  color: gray;
-  display: block;
-}
-.tdContainer-1 {
-  font-size: 14px;
-}
-.tdContainer-2 {
-  font-size: 14px;
-}
-.startButton {
-  position: absolute;
-  top: 48%;
-  right: 0px;
-  width: 67px;
-  height: 40px;
-  transform: translateY(-50%);
-  text-transform: uppercase;
-  font-size: 12px;
-}
-.bigContainer {
-  display: flex;
-  width: 75%;
-  justify-content: space-between;
-  align-items: center;
-}
